taskit-backend/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
behave==1.2.6
behave-django==1.4.0
black==22.1.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
mypy-extensions==0.4.3
parse==1.19.0
parse-type==0.6.0
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
pycparser==2.21
pyhamcrest==2.0.3
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tomli==2.0.0
typing-extensions==4.0.1
whitenoise==5.3.0
